+++ source/example/F0.Cli.Example/Http/INuGetClient.cs
@@ -7,5 +7,6 @@ namespace F0.Cli.Example.Http
 	{
 		Task<string> GetByOwnerAsync(string owner, CancellationToken cancellationToken);
 		Task<string> GetByIdAsync(string id, CancellationToken cancellationToken);
+		Task<string> GetByTagAsync(string tag, int skip, int take, CancellationToken cancellationToken);
 	}
 }