+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Added option binding of _built-in numeric types (C#)_ and `System.Numerics.BigInteger`.
 
 ## v0.4.0 (2019-12-31)
 - Updated dependency on Generic Host from `Microsoft.Extensions.Hosting 3.0.0` to `Microsoft.Extensions.Hosting 3.1.0`.